Completely dissolve tests_general.py
