+++ evap/evaluation/tests/test_coverage.py
@@ -1,217 +1,21 @@
-from django.core.urlresolvers import reverse
-from django.test import TestCase
 from django.test.utils import override_settings
 from django.forms.models import inlineformset_factory
 from django.core import mail
-from django.core.management import call_command
-from django.conf import settings
-from django.contrib.auth.models import Group
 
-from evap.evaluation.models import Semester, Questionnaire, Question, UserProfile, Course, \
-                            Contribution, EmailTemplate, Degree, CourseType
+from evap.evaluation.models import Semester, Questionnaire, UserProfile, Course, \
+                            EmailTemplate, Degree, CourseType, Contribution
 from evap.evaluation.tests.test_utils import WebTest, lastform
 from evap.staff.forms import ContributionFormSet, ContributionForm
-from evap.staff.tools import merge_users
-from evap.rewards.models import RewardPointGranting, RewardPointRedemption
 
 from model_mommy import mommy
-from webtest import AppError
-
-import os.path
 
 
-@override_settings(INSTITUTION_EMAIL_DOMAINS=["institution.com", "student.institution.com"])
-class SampleXlsTests(WebTest):
 
-    @classmethod
-    def setUpTestData(cls):
-        mommy.make(Semester, pk=1)
-        mommy.make(UserProfile, username="user", groups=[Group.objects.get(name="Staff")])
-        mommy.make(CourseType, name_de="Vorlesung", name_en="Vorlesung")
-        mommy.make(CourseType, name_de="Seminar", name_en="Seminar")
-
-    def test_sample_xls(self):
-        page = self.app.get("/staff/semester/1/import", user='user')
-
-        original_user_count = UserProfile.objects.count()
-
-        form = lastform(page)
-        form["vote_start_date"] = "2015-01-01"
-        form["vote_end_date"] = "2099-01-01"
-        form["excel_file"] = (os.path.join(settings.BASE_DIR, "static", "sample.xls"),)
-        form.submit(name="operation", value="import")
-
-        self.assertEqual(UserProfile.objects.count(), original_user_count + 4)
-
-    def test_sample_user_xls(self):
-        page = self.app.get("/staff/user/import", user='user')
-
-        original_user_count = UserProfile.objects.count()
-
-        form = lastform(page)
-        form["excel_file"] = (os.path.join(settings.BASE_DIR, "static", "sample_user.xls"),)
-        form.submit(name="operation", value="import")
-
-        self.assertEqual(UserProfile.objects.count(), original_user_count + 2)
-
-
-class UsecaseTests(WebTest):
-
-    @classmethod
-    def setUpTestData(cls):
-        mommy.make(UserProfile, username="staff.user", groups=[Group.objects.get(name="Staff")])
-        mommy.make(CourseType, name_de="Vorlesung", name_en="Vorlesung")
-        mommy.make(CourseType, name_de="Seminar", name_en="Seminar")
-
-    def test_import(self):
-        page = self.app.get(reverse("staff:index"), user='staff.user')
-
-        # create a new semester
-        page = page.click("[Cc]reate [Nn]ew [Ss]emester")
-        semester_form = lastform(page)
-        semester_form['name_de'] = "Testsemester"
-        semester_form['name_en'] = "test semester"
-        page = semester_form.submit().follow()
-
-        # retrieve new semester
-        semester = Semester.objects.get(name_de="Testsemester",
-                                        name_en="test semester")
-
-        self.assertEqual(semester.course_set.count(), 0, "New semester is not empty.")
-
-        # save original user count
-        original_user_count = UserProfile.objects.count()
-
-        # import excel file
-        page = page.click("[Ii]mport")
-        upload_form = lastform(page)
-        upload_form['vote_start_date'] = "02/29/2000"
-        upload_form['vote_end_date'] = "02/29/2012"
-        upload_form['excel_file'] = (os.path.join(settings.BASE_DIR, "staff/fixtures/test_enrolment_data.xls"),)
-        upload_form.submit(name="operation", value="import").follow()
-
-        self.assertEqual(UserProfile.objects.count(), original_user_count + 23)
-
-        courses = Course.objects.filter(semester=semester).all()
-        self.assertEqual(len(courses), 23)
-
-        for course in courses:
-            responsibles_count = Contribution.objects.filter(course=course, responsible=True).count()
-            self.assertEqual(responsibles_count, 1)
-
-        check_student = UserProfile.objects.get(username="diam.synephebos")
-        self.assertEqual(check_student.first_name, "Diam")
-        self.assertEqual(check_student.email, "diam.synephebos@student.hpi.uni-potsdam.de")
-
-        check_contributor = UserProfile.objects.get(username="sanctus.aliquyam.ext")
-        self.assertEqual(check_contributor.first_name, "Sanctus")
-        self.assertEqual(check_contributor.last_name, "Aliquyam")
-        self.assertEqual(check_contributor.email, "567@web.de")
-
-    def test_login_key(self):
-        self.assertRedirects(self.app.get(reverse("results:index")), "/?next=/results/")
-
-        user = mommy.make(UserProfile)
-        user.generate_login_key()
-
-        url_with_key = reverse("results:index") + "?userkey=%s" % user.login_key
-        self.app.get(url_with_key)
-
-    def test_create_questionnaire(self):
-        page = self.app.get(reverse("staff:index"), user="staff.user")
-
-        # create a new questionnaire
-        page = page.click("[Cc]reate [Nn]ew [Qq]uestionnaire")
-        questionnaire_form = lastform(page)
-        questionnaire_form['name_de'] = "Test Fragebogen"
-        questionnaire_form['name_en'] = "test questionnaire"
-        questionnaire_form['public_name_de'] = "Oeffentlicher Test Fragebogen"
-        questionnaire_form['public_name_en'] = "Public Test Questionnaire"
-        questionnaire_form['question_set-0-text_de'] = "Frage 1"
-        questionnaire_form['question_set-0-text_en'] = "Question 1"
-        questionnaire_form['question_set-0-type'] = "T"
-        questionnaire_form['index'] = 0
-        questionnaire_form.submit().follow()
-
-        # retrieve new questionnaire
-        questionnaire = Questionnaire.objects.get(name_de="Test Fragebogen", name_en="test questionnaire")
-        self.assertEqual(questionnaire.question_set.count(), 1, "New questionnaire is empty.")
-
-    def test_create_empty_questionnaire(self):
-        page = self.app.get(reverse("staff:index"), user="staff.user")
-
-        # create a new questionnaire
-        page = page.click("[Cc]reate [Nn]ew [Qq]uestionnaire")
-        questionnaire_form = lastform(page)
-        questionnaire_form['name_de'] = "Test Fragebogen"
-        questionnaire_form['name_en'] = "test questionnaire"
-        questionnaire_form['public_name_de'] = "Oeffentlicher Test Fragebogen"
-        questionnaire_form['public_name_en'] = "Public Test Questionnaire"
-        questionnaire_form['index'] = 0
-        page = questionnaire_form.submit()
-
-        self.assertIn("You must have at least one of these", page)
-
-        # retrieve new questionnaire
-        with self.assertRaises(Questionnaire.DoesNotExist):
-            Questionnaire.objects.get(name_de="Test Fragebogen", name_en="test questionnaire")
-
-    def test_copy_questionnaire(self):
-        questionnaire = mommy.make(Questionnaire, name_en="Seminar")
-        mommy.make(Question, questionnaire=questionnaire)
-        page = self.app.get(reverse("staff:index"), user="staff.user")
-
-        # create a new questionnaire
-        page = page.click("All questionnaires")
-        page = page.click("Copy", index=1)
-        questionnaire_form = lastform(page)
-        questionnaire_form['name_de'] = "Test Fragebogen (kopiert)"
-        questionnaire_form['name_en'] = "test questionnaire (copied)"
-        questionnaire_form['public_name_de'] = "Oeffentlicher Test Fragebogen (kopiert)"
-        questionnaire_form['public_name_en'] = "Public Test Questionnaire (copied)"
-        page = questionnaire_form.submit().follow()
-
-        # retrieve new questionnaire
-        questionnaire = Questionnaire.objects.get(name_de="Test Fragebogen (kopiert)", name_en="test questionnaire (copied)")
-        self.assertEqual(questionnaire.question_set.count(), 1, "New questionnaire is empty.")
-
-    def test_assign_questionnaires(self):
-        semester = mommy.make(Semester, name_en="Semester 1")
-        mommy.make(Course, semester=semester, type=CourseType.objects.get(name_de="Seminar"), contributions=[
-            mommy.make(Contribution, contributor=mommy.make(UserProfile),
-                       responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS)])
-        mommy.make(Course, semester=semester, type=CourseType.objects.get(name_de="Vorlesung"), contributions=[
-            mommy.make(Contribution, contributor=mommy.make(UserProfile),
-                       responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS)])
-        questionnaire = mommy.make(Questionnaire)
-        page = self.app.get(reverse("staff:index"), user="staff.user")
-
-        # assign questionnaire to courses
-        page = page.click("Semester 1", index=0)
-        page = page.click("Assign Questionnaires")
-        assign_form = lastform(page)
-        assign_form['Seminar'] = [questionnaire.pk]
-        assign_form['Vorlesung'] = [questionnaire.pk]
-        page = assign_form.submit().follow()
-
-        for course in semester.course_set.all():
-            self.assertEqual(course.general_contribution.questionnaires.count(), 1)
-            self.assertEqual(course.general_contribution.questionnaires.get(), questionnaire)
-
-    def test_remove_responsibility(self):
-        user = mommy.make(UserProfile)
-        contribution = mommy.make(Contribution, contributor=user, responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS)
-
-        page = self.app.get(reverse("staff:index"), user="staff.user")
-        page = page.click(contribution.course.semester.name_en, index=0)
-        page = page.click(contribution.course.name_en)
-
-        # remove responsibility
-        form = lastform(page)
-        form['contributions-0-responsibility'] = "CONTRIBUTOR"
-        page = form.submit()
-
-        self.assertIn("No responsible contributor found", page)
+"""
+These tests were created to get a higher test coverage. Some actually contain functional
+tests and should be moved to their appropriate place, others don't really test anything
+and should be replaced by better tests. Eventually, this file is to be removed.
+"""
 
 
 @override_settings(INSTITUTION_EMAIL_DOMAINS=["example.com"])
@@ -299,17 +103,6 @@ class URLTests(WebTest):
         self.get_assert_403("/student/vote/5", "student")
         self.get_assert_403("/results/semester/1/course/8", "student"),
         self.get_assert_403("/results/semester/1/course/7", "student"),
-
-    def test_redirecting_urls(self):
-        """
-            Tests whether some pages that cannot be accessed (e.g. for courses in certain states)
-            do not return 200 but redirect somewhere else.
-        """
-        tests = [
-        ]
-
-        for _, url, user in tests:
-            self.get_assert_302(url, user)
 
     def test_failing_forms(self):
         """
@@ -669,174 +462,3 @@ class URLTests(WebTest):
         self.assertEqual(Course.objects.filter(type=main_type).count(), num_courses_with_main_type + 1)
         for course in courses_with_other_type:
             self.assertTrue(course.type == main_type)
-
-
-class TestDataTest(TestCase):
-
-    def load_test_data(self):
-        """
-            Asserts that the test data still load cleanly.
-            This test does not have the "test_" prefix, as it is meant
-            to be started manually e.g. by Travis.
-        """
-        try:
-            call_command("loaddata", "test_data", verbosity=0)
-        except Exception:
-            self.fail("Test data failed to load.")
-
-
-class MergeUsersTest(TestCase):
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.user1 = mommy.make(UserProfile, username="test1")
-        cls.user2 = mommy.make(UserProfile, username="test2")
-        cls.user3 = mommy.make(UserProfile, username="test3")
-        cls.group1 = mommy.make(Group, name="group1")
-        cls.group2 = mommy.make(Group, name="group2")
-        cls.main_user = mommy.make(UserProfile,
-            username="main_user",
-            title="Dr.",
-            first_name="Main",
-            last_name="",
-            email="",  # test that merging works when taking the email from other user (UniqueConstraint)
-            groups=[cls.group1],
-            delegates=[cls.user1, cls.user2],
-            represented_users=[cls.user3],
-            cc_users=[cls.user1],
-            ccing_users=[]
-        )
-        cls.other_user = mommy.make(UserProfile,
-            username="other_user",
-            title="",
-            first_name="Other",
-            last_name="User",
-            email="other@test.com",
-            groups=[cls.group2],
-            delegates=[cls.user3],
-            represented_users=[cls.user1],
-            cc_users=[],
-            ccing_users=[cls.user1, cls.user2],
-            is_superuser=True
-        )
-        cls.course1 = mommy.make(Course, name="course1", participants=[cls.main_user, cls.other_user])  # this should make the merge fail
-        cls.course2 = mommy.make(Course, name="course2", participants=[cls.main_user], voters=[cls.main_user])
-        cls.course3 = mommy.make(Course, name="course3", participants=[cls.other_user], voters=[cls.other_user])
-        cls.contribution1 = mommy.make(Contribution, contributor=cls.main_user, course=cls.course1)
-        cls.contribution2 = mommy.make(Contribution, contributor=cls.other_user, course=cls.course1)  # this should make the merge fail
-        cls.contribution3 = mommy.make(Contribution, contributor=cls.other_user, course=cls.course2)
-        cls.rewardpointgranting_main = mommy.make(RewardPointGranting, user_profile=cls.main_user)
-        cls.rewardpointgranting_other = mommy.make(RewardPointGranting, user_profile=cls.other_user)
-        cls.rewardpointredemption_main = mommy.make(RewardPointRedemption, user_profile=cls.main_user)
-        cls.rewardpointredemption_other = mommy.make(RewardPointRedemption, user_profile=cls.other_user)
-
-    def test_merge_handles_all_attributes(self):
-        user1 = mommy.make(UserProfile)
-        user2 = mommy.make(UserProfile)
-
-        all_attrs = list(field.name for field in UserProfile._meta.get_fields(include_hidden=True))
-
-        # these are relations to intermediate models generated by django for m2m relations.
-        # we can safely ignore these since the "normal" fields of the m2m relations are present as well.
-        all_attrs = list(attr for attr in all_attrs if not attr.startswith("UserProfile_"))
-
-        # equally named fields are not supported, sorry
-        self.assertEqual(len(all_attrs), len(set(all_attrs)))
-
-        # some attributes we don't care about when merging
-        ignored_attrs = {
-            'id',  # nothing to merge here
-            'password',  # not used in production
-            'last_login',  # something to really not care about
-            'user_permissions',  # we don't use permissions
-            'logentry',  # wtf
-            'login_key',  # we decided to discard other_user's login key
-            'login_key_valid_until',  # not worth dealing with
-            'Course_voters+',  # some more intermediate models, for an explanation see above
-            'Course_participants+',  # intermediate model
-        }
-        expected_attrs = set(all_attrs) - ignored_attrs
-
-        # actual merge happens here
-        merged_user, errors, warnings = merge_users(user1, user2)
-        handled_attrs = set(merged_user.keys())
-
-        # attributes that are handled in the merge method but that are not present in the merged_user dict
-        # add attributes here only if you're actually dealing with them in merge_users().
-        additional_handled_attrs = {
-            'grades_last_modified_user+',
-            'course_last_modified_user+',
-        }
-
-        actual_attrs = handled_attrs | additional_handled_attrs
-
-        self.assertEqual(expected_attrs, actual_attrs)
-
-    def test_merge_users(self):
-        merged_user, errors, warnings = merge_users(self.main_user, self.other_user)  # merge should fail
-        self.assertSequenceEqual(errors, ['contributions', 'courses_participating_in'])
-        self.assertSequenceEqual(warnings, ['rewards'])
-
-        # assert that nothing has changed
-        self.main_user.refresh_from_db()
-        self.other_user.refresh_from_db()
-        self.assertEqual(self.main_user.username, "main_user")
-        self.assertEqual(self.main_user.title, "Dr.")
-        self.assertEqual(self.main_user.first_name, "Main")
-        self.assertEqual(self.main_user.last_name, "")
-        self.assertEqual(self.main_user.email, "")
-        self.assertSequenceEqual(self.main_user.groups.all(), [self.group1])
-        self.assertSequenceEqual(self.main_user.delegates.all(), [self.user1, self.user2])
-        self.assertSequenceEqual(self.main_user.represented_users.all(), [self.user3])
-        self.assertSequenceEqual(self.main_user.cc_users.all(), [self.user1])
-        self.assertSequenceEqual(self.main_user.ccing_users.all(), [])
-        self.assertFalse(self.main_user.is_superuser)
-        self.assertTrue(RewardPointGranting.objects.filter(user_profile=self.main_user).exists())
-        self.assertTrue(RewardPointRedemption.objects.filter(user_profile=self.main_user).exists())
-        self.assertEqual(self.other_user.username, "other_user")
-        self.assertEqual(self.other_user.title, "")
-        self.assertEqual(self.other_user.first_name, "Other")
-        self.assertEqual(self.other_user.last_name, "User")
-        self.assertEqual(self.other_user.email, "other@test.com")
-        self.assertSequenceEqual(self.other_user.groups.all(), [self.group2])
-        self.assertSequenceEqual(self.other_user.delegates.all(), [self.user3])
-        self.assertSequenceEqual(self.other_user.represented_users.all(), [self.user1])
-        self.assertSequenceEqual(self.other_user.cc_users.all(), [])
-        self.assertSequenceEqual(self.other_user.ccing_users.all(), [self.user1, self.user2])
-        self.assertSequenceEqual(self.course1.participants.all(), [self.main_user, self.other_user])
-        self.assertSequenceEqual(self.course2.participants.all(), [self.main_user])
-        self.assertSequenceEqual(self.course2.voters.all(), [self.main_user])
-        self.assertSequenceEqual(self.course3.participants.all(), [self.other_user])
-        self.assertSequenceEqual(self.course3.voters.all(), [self.other_user])
-        self.assertTrue(RewardPointGranting.objects.filter(user_profile=self.other_user).exists())
-        self.assertTrue(RewardPointRedemption.objects.filter(user_profile=self.other_user).exists())
-
-        # fix data
-        self.course1.participants = [self.main_user]
-        self.contribution2.delete()
-
-        merged_user, errors, warnings = merge_users(self.main_user, self.other_user)  # merge should succeed
-        self.assertEqual(errors, [])
-        self.assertSequenceEqual(warnings, ['rewards']) # rewards warning is still there
-
-        self.main_user.refresh_from_db()
-        self.assertEqual(self.main_user.username, "main_user")
-        self.assertEqual(self.main_user.title, "Dr.")
-        self.assertEqual(self.main_user.first_name, "Main")
-        self.assertEqual(self.main_user.last_name, "User")
-        self.assertEqual(self.main_user.email, "other@test.com")
-        self.assertSequenceEqual(self.main_user.groups.all(), [self.group1, self.group2])
-        self.assertSequenceEqual(self.main_user.delegates.all(), [self.user1, self.user2, self.user3])
-        self.assertSequenceEqual(self.main_user.represented_users.all(), [self.user1, self.user3])
-        self.assertSequenceEqual(self.main_user.cc_users.all(), [self.user1])
-        self.assertSequenceEqual(self.main_user.ccing_users.all(), [self.user1, self.user2])
-        self.assertSequenceEqual(self.course1.participants.all(), [self.main_user])
-        self.assertSequenceEqual(self.course2.participants.all(), [self.main_user])
-        self.assertSequenceEqual(self.course2.voters.all(), [self.main_user])
-        self.assertSequenceEqual(self.course3.participants.all(), [self.main_user])
-        self.assertSequenceEqual(self.course3.voters.all(), [self.main_user])
-        self.assertTrue(self.main_user.is_superuser)
-        self.assertTrue(RewardPointGranting.objects.filter(user_profile=self.main_user).exists())
-        self.assertTrue(RewardPointRedemption.objects.filter(user_profile=self.main_user).exists())
-        self.assertFalse(RewardPointGranting.objects.filter(user_profile=self.other_user).exists())
-        self.assertFalse(RewardPointRedemption.objects.filter(user_profile=self.other_user).exists())
